+++ CHANGELOG.md
@@ -5,6 +5,22 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## 0.1.1-beta - 2022-10-24
+
+### Added
+
+- HTTPS/TLS support (#4)
+
+
+### Fixed
+
+- Default params für Discovery ergänzt (#3)
+- Custom HTTP port übergabe gefixt (#1)
+
+### Changed
+
+- VPN-Status Zeitstempel: geändert auf "letzte Statusänderung"
+
 ## 0.1-beta - 2022-10-21
 
 Initial release